Add file agent example as a service on flusk TODO: dockerize
